🗝️🗝️ Adding isBlocking static function for YubiKey
Nothing was logged on our side when the yk initialization
failed. This patch will add logging and detailed error
message (if possible) in the case of a failing yubikey
interaction, whether before of after initialization.

The patch also extracts a static `isBlocking` function, which
easily can be easily reusable from other components of the app
(like, you know, the CLI 😉😉😎😎)

also a typo in the yk test files 🖊️🖊️